Reorganized generators folder
Main files are now separated from additional lib files.
