+++ app/ascii_generators/txt2ascii_lib/art_dic.py
