+++ app/ascii_generators/txt2ascii_lib/text_dic1.py
