+++ app/ascii_generators/txt2ascii_lib/text_dic2.py
