+++ app/ascii_generators/txt2ascii_lib/text_dic3.py
